# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DFDR Swagger OpenAPI Documentation",
        "description": "Endpoint from DFDR Application",
        "contact": {
            "email": "admin@dfdr.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/category": {
            "get": {
                "tags": [
                    "Category"
                ],
                "description": "View All",
                "operationId": "findAllCategory",
                "responses": {
                    "default": {
                        "description": "Success get data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "description": "Add new",
                "operationId": "postCategory",
                "requestBody": {
                    "description": "Add new Category",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_name"
                                ],
                                "properties": {
                                    "category_name": {
                                        "description": "Category Name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "description": "View By Id",
                "operationId": "showCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Input Data"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "description": "Update Data",
                "operationId": "putCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Success Update Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "description": "Delete Data",
                "operationId": "destroyCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of Category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Delete Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category/search/{name}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "description": "Search Data",
                "operationId": "searchCategory",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Name of Category",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Success Search Data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid Parameter"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer -token from login-)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}